Added AbstractReplay and made IBXM/Micromod extend it
